Working on Flask Functionality
__init__
routes
main app
