+++ WriteMyRights.py
@@ -1,0 +1,10 @@
+from app import app
+# from app import app, getApp
+#
+#
+# @app.shell_context_processor
+# def make_shell_context():
+#     return None
+#
+#
+# app = getApp()
