+++ app/__init__.py
@@ -1,0 +1,10 @@
+from flask import Flask
+
+app = Flask(__name__)
+
+from app import routes
+# def getApp():
+#     return app
+
+
+
