+++ app/routes.py
@@ -1,0 +1,8 @@
+from app import app
+from flask import render_template
+
+@app.route('/')
+@app.route('/index')
+def index():
+    return "Hello, World!"
+    #return render_template("index.html", title='Write My Rights')
