--- index.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <title></title>
-        <meta charset = 'UTF8'>
-        <script></script>
-    </head>
-    <body>
-        <h1>Write My Rights</h1>
-        <p>Sample Text</p>
-    </body>
-</html>
